Yeah

+++ CapstoneBackend.Core/Services/MainService.cs
@@ -17,19 +17,29 @@ public class MainService
 
     public Main CreateEntry(Main main)
     {
+        if (main.Id != 0)
+        {
+            throw new ArgumentException("Id must be 0 for new entries.", nameof(main.Id));
+        }
         return _mainRepository.CreateEntry(main);
     }
 
     public Main UpdateEntry(Main main)
     {
-        if(main.Id == 0) throw new Exception("Can't be 0!");
-        _ = _mainRepository.GetEntryById(main.Id) ?? throw new Exception("No entry found.");
+        
+        if(main.Id <= 0) throw new ArgumentOutOfRangeException(nameof(main.Id), "Id must be greater than 0.");
+        var _ = _mainRepository.GetEntryById(main.Id) ?? 
+                throw new KeyNotFoundException($"No entry found with 'Id': '{main.Id}'.");
         return _mainRepository.UpdateEntry(main);
     }
 
     public Main GetEntryById(int id)
     {
-        var result = _mainRepository.GetEntryById(id) ?? throw new Exception("No entry found.");
+        if (id <= 0)
+        {
+            throw new ArgumentOutOfRangeException(nameof(id), "Id must be greater than 0.");
+        }
+        var result = _mainRepository.GetEntryById(id) ?? throw new KeyNotFoundException($"No entry found with 'id' {id}.");
         return result;
     }
 
@@ -41,7 +51,8 @@ public class MainService
 
     public bool DeleteEntryById(int id)
     {
-        _ = _mainRepository.GetEntryById(id) ?? throw new Exception("No Entry Found");
+        if(id <= 0) throw new ArgumentOutOfRangeException(nameof(id), "Id must be greater than 0.");
+        _ = _mainRepository.GetEntryById(id) ?? throw new KeyNotFoundException("No Entry Found");
         return _mainRepository.DeleteEntryById(id);
     }
